requirements.txt
django
django-storages[azure]
django-scheduler
djangorestframework
djangorestframework-jsonapi
markdown       
django-filter